Added listeners for entrance activity + minor folder refactoring
