+++ app/src/main/java/com/example/vinilos/ui/guest/GuestViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.vinilos
+package com.example.vinilos.ui.guest
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
