+++ app/src/main/res/layout/activity_entrance.xml
@@ -31,7 +31,6 @@
         android:elevation="5dp"
         android:insetLeft="20dp"
         android:insetRight="20dp"
-        android:onClick="onCollectorButtonClick"
         android:padding="20dp"
         android:text="@string/collector"
         android:textSize="20sp"
@@ -56,7 +55,6 @@
         android:elevation="5dp"
         android:insetLeft="20dp"
         android:insetRight="20dp"
-        android:onClick="onGuestButtonClick"
         android:padding="20dp"
         android:text="@string/guest"
         android:textSize="20sp"
